Restore preoutput filters from before when turning mode off
Turning the mode on stores the `comint-preoutput-filter-functions` in the
variable `stfu-process--original-preoutput-filters`. Turning the mode off
causes this value to be restored to `comint-preoutput-filter-functions`.

In the future, this may get changed to just remove the custom
preoutput filter from `comint-preoutput-filter-functions`. This would
prevent the potential downside of clobbering any preoutput filters
that are added after this mode is enabled.