Add stfu-now to key map; fix bug when turning off minor mode
Also add advice about setting stfu-now in global keymap to initj

The `define-minor-mode` had erroneously checked to see if the
`stfu-process-mode-map` was truthy or nil when it was supposed to
check `stfu-process-map` instead. This was likely the result of
errant code completion.

Ths update also renames the mode map to have the proper prefix.